requirements.txt
google-generativeai
pdfplumber
pypandoc
edge-tts
openai
python-dotenv
python-docx
requests
